requirements.txt
Flask==3.1.0
PyMySQL==1.1.2
bcrypt==5.0.0
pytesseract==0.3.13
Pillow==11.0.0
opencv-python==4.12.0.88
numpy==2.1.2
spacy==3.8.11
sentence-transformers==5.2.3
faiss-cpu==1.13.2
python-dotenv==1.1.0
